MAGECLOUD-946: Cover with Unit tests
